Add files via upload
Previous UI has had several bugs (see below) and could not be run on Http-server. In this version, the bugs have been addressed and it works with http-server. It has been tested.

issues with the previous version: 


1- owner could not register a client--- reason: myContract.register_client(client_addr...)
was commneted out

2- insert age couldn't be called as insert_age() did not have "hash_of_file" in the argument-- it was not compatible with the most recent version of the contract. I defined:
var hash_of_file = 0x111;   and passed hash_of_file to insert_age() too.  In the local UI we need to allow a client to upload a file, then the UI would hash the file. Next the UI "encrypts" that hash, and displaies it on the screen. Later on the client can insert the encryption when it's inserting it encryption of the age in the server side UI.  The same procedure is done for insert_degree() and insert_license(). 



3- 95% of "current accounts" were set to user.account instead of account. 
4- the owner page for registration was not working as all CCS, image, .js files were in a directorry undefiend-- I moved all to the same directory as main.
5- in validatro forlder--> index.html, on line 170-187, there was something like:

      $(function () {
        var scntDiv = $('#p_scents');
        var i = $('#p_scents p').size() + 1;

        $('#addScnt').live('click', function () {
            $('<p><label for="p_scnts"><input type="text" id="p_scnt" size="20" name="p_scnt_' + i + '" value="" placeholder="Input Value" /></label> <a href="#" id="remScnt">Remove</a></p>').appendTo(scntDiv);
            i++;
            return false;
        });

        $('#remScnt').live('click', function () {
            if (i > 2) {
                $(this).parents('p').remove();
                i--;
            }
            return false;
        });
    });

That coused a problem. I have commented them out. 

6- hashage in insert_age() was not defined- so the browser was alerting about it. I had to comment out lines 1426-1428